Refactor tests with .mocharc.js and root hooks
