+++ .mocharc.js
@@ -1,0 +1,7 @@
+'use strict'
+
+module.exports = {
+  exit: true,
+  timeout: 30000,
+  file: ['./test/setup_browser.js']
+};
