+++ test/integration/user_downvotes_page.test.js
@@ -1,40 +1,12 @@
-const path = require('path');
-const puppeteer = require('puppeteer');
 const assert = require("assert");
-const extensionPath = path.resolve();
+const { getExtensionId } = require('../integrationTestHelper');
 
-async function bootBrowser() {
-  return puppeteer.launch({
-    headless: false,
-    args: [
-      `--disable-extensions-except=${extensionPath}`,
-      `--load-extension=${extensionPath}`,
-      '--no-sandbox',
-      '--disable-setuid-sandbox',
-    ],
-  });
-}
-
-describe('User upvotes page', function () {
-  let browser;
+describe('User downvotes page', function () {
   let page;
-  this.timeout(20000); // default is 2 seconds and that may not be enough to boot browsers and pages.
-
-  before(async () => {
-    browser = await bootBrowser();
-  });
-
-  after(async () => {
-    await browser.close();
-  });
 
   beforeEach(async () => {
     page = await browser.newPage();
-    await page.goto('chrome://extensions');
-    const extensionId = await page.evaluate(async () => {
-      const [ext] = await chrome.management.getAll();
-      return ext.id;
-    });
+    const extensionId = await getExtensionId(page);
     await page.goto(`chrome-extension://${extensionId}/dialog.html`);
   });
 
@@ -43,7 +15,7 @@ describe('User upvotes page', function () {
   });
 
   context('when popup.html', () => {
-    it('clicks upvote button', async () => {
+    it('clicks downvote button', async () => {
       await page.setRequestInterception(true)
       page.on('request', (request) => {
         if (request.url() === 'http://localhost:8000/api/v1/source_articles/submit/') {
@@ -71,7 +43,7 @@ describe('User upvotes page', function () {
         }
       });
 
-      await page.click('button[name="upvote"]');
+      await page.click('button[name="downvote"]');
       await page.waitForSelector('.r-response');
       const responseText = await page.$eval('#r-response', response => response.innerText);
       assert.strictEqual(responseText, 'POŽIADAVKA BOLA ODOSLANÁ.\nĎAKUJEME');
